Rewriting calculate changes functions to typescript
